forcegps: Fix bug related to exiftool warnings
Any time a file resulted in an exiftool warning, the script was ignoring
that file. This change tells the exiftool calls to ignore minor errors
and warnings so that the files will be updated correctly.

I could still see the script struggling with files that have existing
EXIF issues that exiftool doesn't handle well, but since this is a HACK
I'm not going to lose sleep over it. ¯\_(ツ)_/¯